Render json from payment methods
